AMP-2880 - package.json: upgrade dependencies to latest within the limit of compatibility with current Vue 2 version, while supported by upgraded Node version - tmp fix for css related errors (as mentioned in AMP-1811) - temporarily surppress eslint errors to allow compilation (upgraded eslint has stricter requirements that would break current code) - rename Header to AmpHeader as required by eslint - tmp removal of bbs editor related code to allow compilation
