+++ src/components/collections/CollectionDetails.vue
@@ -128,5 +128,5 @@ export default {
 };
 </script>
 <style scoped>
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 </style>
