+++ src/components/collections/CollectionDetailsItemCard.vue
@@ -188,5 +188,5 @@ export default {
 };
 </script>
 <style scoped>
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 </style>
