+++ src/components/collections/Collections.vue
@@ -151,5 +151,5 @@ export default {
 </script>
 
 <style scoped>
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 </style>
