+++ src/components/evaluation/GroundTruthModal.vue
@@ -385,7 +385,7 @@ export default {
 </script>
 
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 .body-m {
   min-height: 32rem;
 }
