+++ src/components/evaluation/MGMevaluation.vue
@@ -248,7 +248,7 @@ export default {
 </script>
 
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 main {
   margin-top: 0px;
 }
