+++ src/components/hmgm/TokenValidator.vue
@@ -28,7 +28,7 @@
 </template>
 
 <script>
-import AmpHeader from "@/components/shared/Header.vue";
+import AmpHeader from "@/components/shared/AmpHeader.vue";
 import Logout from "@/components/shared/Logout.vue";
 import Modal from "@/components/shared/Modal.vue";
 import { auth_token_required, auth_token_valid } from "@/service/hmgm-service";
