+++ src/components/hmgm/TranscriptEditor.vue
@@ -28,6 +28,7 @@
             />
           </div>
         </div>
+        <!--
         <BBCTranscriptEditor
           :key="key"
           v-if="transcriptDataValue && sttType"
@@ -41,6 +42,7 @@
           :ref="player"
         >
         </BBCTranscriptEditor>
+      -->
       </div>
     </div>
     <modal v-if="showModal" @close="modalDismiss" class="my-modal">
@@ -123,11 +125,11 @@
 </template>
 
 <script>
-import AmpHeader from "@/components/shared/Header.vue";
+import AmpHeader from "@/components/shared/AmpHeader.vue";
 import Loader from "@/components/shared/Loader.vue";
 import Logout from "@/components/shared/Logout.vue";
 import TokenValidator from "@/components/hmgm/TokenValidator";
-import BBCTranscriptEditor from "@bbc/react-transcript-editor/dist";
+// import BBCTranscriptEditor from "@bbc/react-transcript-editor/dist";
 import Modal from "@/components/shared/Modal.vue";
 import {
   getTranscript,
@@ -142,7 +144,7 @@ export default {
     AmpHeader,
     Loader,
     Logout,
-    BBCTranscriptEditor,
+    // BBCTranscriptEditor,
     TokenValidator,
     Modal,
   },
