+++ src/components/home/index.vue
@@ -71,7 +71,7 @@ export default {
 </script>
 
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 html {
   height: 100%;
 }
