+++ src/components/supplement/SupplementList.vue
@@ -273,7 +273,7 @@ export default {
 </script>
 
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 
 .btn-grp {
   margin-bottom: 33px !important;
