+++ src/components/unused/WorkflowFiles.vue
@@ -62,5 +62,5 @@ export default {
 </script>
 
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 </style>
