+++ src/components/workflow/SaveBundle.vue
@@ -253,7 +253,7 @@ export default {
 </script>
 
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 
 .primary-button {
   float: right;
