+++ src/components/workflow/SelectBundle.vue
@@ -264,7 +264,7 @@ export default {
 }
 </style>
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 .bundles {
   width: 100%;
 }
