+++ src/components/workflow/SelectWorkflow.vue
@@ -909,7 +909,7 @@ export default {
 </script>
 
 <style lang="css">
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 .node-li {
   margin-bottom: 16px;
 }
