+++ src/components/workflow/WorkflowEditor.vue
@@ -56,7 +56,7 @@
 </template>
 
 <script>
-import AmpHeader from "@/components/shared/Header.vue";
+import AmpHeader from "@/components/shared/AmpHeader.vue";
 import Logout from "@/components/shared/Logout.vue";
 import Modal from "@/components/shared/Modal.vue";
 import { env } from "../../helpers/env";
