+++ src/components/workflow/WorkflowSubmission.vue
@@ -50,7 +50,7 @@ export default {
 };
 </script>
 <style scoped>
-@import "/amppd-ui/src/styles/style.css";
+@import "../../styles/style.css";
 .main-margin-min {
   margin: 0.5em !important;
 }
